+++ src/main/java/blackjack/dto/TotalResultDTO.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.dto;
 
 import java.util.List;
